+++ zfit/models/dist_tfp.py
@@ -200,7 +200,7 @@ class Gauss(WrapDistribution):
         params = OrderedDict((('mu', mu), ('sigma', sigma)))
         dist_params = dict(loc=mu, scale=sigma)
         distribution = tfp.distributions.Normal
-        super().__init__(distribution=distribution, dist_params=dist_params, obs=obs, params=params, name=name + "_tfp")
+        super().__init__(distribution=distribution, dist_params=dist_params, obs=obs, params=params, name=name)
 
 
 class ExponentialTFP(WrapDistribution):
@@ -211,7 +211,7 @@ class ExponentialTFP(WrapDistribution):
         params = OrderedDict((('tau', tau),))
         dist_params = dict(rate=tau)
         distribution = tfp.distributions.Exponential
-        super().__init__(distribution=distribution, dist_params=dist_params, obs=obs, params=params, name=name + "_tfp")
+        super().__init__(distribution=distribution, dist_params=dist_params, obs=obs, params=params, name=name)
 
 
 class Uniform(WrapDistribution):
@@ -231,7 +231,7 @@ class Uniform(WrapDistribution):
         params = OrderedDict((("low", low), ("high", high)))
         dist_params = dict(low=low, high=high)
         distribution = tfp.distributions.Uniform
-        super().__init__(distribution=distribution, dist_params=dist_params, obs=obs, params=params, name=name + "_tfp")
+        super().__init__(distribution=distribution, dist_params=dist_params, obs=obs, params=params, name=name)
 
 
 class TruncatedGauss(WrapDistribution):